More changes and fixes to pass normals around separately from MVert

+++ source/blender/editors/sculpt_paint/sculpt_intern.h
@@ -231,7 +231,7 @@ typedef struct {
 
   /* Original coordinate, normal, and mask. */
   const float *co;
-  const float *no;
+  const short *no;
   float mask;
   const float *col;
 } SculptOrigVertData;
